requirements.txt
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.14.2
blinker==1.4
cachetools==5.5.2
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
distro==1.7.0
Django==5.2.1
django-extensions==4.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-dialogflow==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.20.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==6.0.2
Markdown==3.9
MarkupSafe==2.0.1
more-itertools==8.10.0
oauthlib==3.2.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.23
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.13.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytz==2022.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
